requirements.txt
allure_python_commons==2.15.0
clickhouse_sqlalchemy==0.3.2
jenkins==1.0.2
jsonpath==0.82.2
pandas==2.3.3
paramiko==4.0.0
pip==25.2
pymongo==4.15.5
pymysql==1.1.2
PyQt5==5.15.11
PyQt5_sip==12.17.0
pytest==8.4.0
pytest==7.4.0
PyYAML==6.0.2
PyYAML==6.0.3
redis==7.1.0
Requests==2.32.5
SQLAlchemy==2.0.45
urllib3==2.6.2
xlrd==2.0.2
xlutils==2.0.0
xlwt==1.3.0

